Skeletal Increment
* Setup Server Polling

This is a test of how we might implement polling. I have some dummy data returned along with an incrementing num on the /game endpoint.

* Send Message and Clear Post Request

Added buttons to send post requests to send messages with username and clear message history.

* Add Make Suggestion

Added a button to make suggestion. Takes the three values from the drop down lists and sends them to the server along with the username.

* Added Make Accusation

Added make accusation button that will demonstrate the server giving specific info to a single client.

* Increased Poll Server Timer

* Trying out different way of polling the server

* Put polling back to how it was

* made accusations and suggestions persist in DB

* Updated Procfile to make migrations