Pregame lobby, models, and views
* created all models, filled out some methods and left some as stubs

* finished defining models

* started creating views

* Initial attempt at pregame lobby

This is a pregame lobby concept for the minimal increment.

* Trying to fix error in getLobbyPlayers

* Fixed getLobbyPlayers

* Updated player list on the client

Taking the values from getLobbyPlayers and putting them in an unordered list on the page

* Fixed leaveLobby Bug

Bug occurs when you log out in another tab and try to leave lobby.

* Redirect Player to lobby

if player was in lobby before closing browser, redirect them the when they open up the base url

* Setup front end create game flow

Need to give actual functionality to create game button soon.

* Finished minimal create game flow

* Guard against null current player and cleanup

* Added redirects to get user to right page

* Added comments and made create and start game happen together for the minimal

* Put in basic game board

* More game page setup

* Added more board game setup and fixed html formatting issues

Atom caused a bunch of stupid formatting issues. I'm using VSCode now.